agent/batch: add support for exception events
